# Faster whitespace skipping