In the existing code, there were many places where ignoreSpaceRun() had
to be called. This did a series of peek()/take() calls and was quite
slow for the same reasons as the old takeString().

This commit make the emit() function always skip whitespace after a
token. The whitespace skipping is implemented in a faster way than using
peek()/take().

In addition to improving performance, this also simplifies the
json_states code because it doesn't have to check for whitespace anymore
(except at the beginning of parsing).

Additional improvement from this commit (relative to only the takeString
optimization):

benchmark                      old ns/op     new ns/op     delta
BenchmarkUnmarshalMix          4472287       4259434       -4.76%
BenchmarkDecodeMix             1594484       1604531       +0.63%
BenchmarkBytesMix              1232586       1044650       -15.25%
BenchmarkReaderMix             3368036       3174426       -5.75%

Cumulative improvement of this commit and takeString optimization
together:

benchmark                      old ns/op     new ns/op     delta
BenchmarkUnmarshalMix          4389407       4259434       -2.96%
BenchmarkDecodeMix             1595274       1604531       +0.58%
BenchmarkBytesMix              1550321       1044650       -32.62%
BenchmarkReaderMix             3878020       3174426       -18.14%